Extract OpenApi integration tests into dedicated project
Moves the OpenAPI document tests out of StrongTypes.Api.IntegrationTests
(which now focuses on EF Core round-tripping) and into a new
StrongTypes.OpenApi.IntegrationTests project that hosts both pipelines'
test apps directly. The Microsoft side now uses the new
StrongTypes.OpenApi.TestApi.Microsoft host instead of the EF-Core-coupled
StrongTypes.Api, so its tests no longer pay SQL/Postgres container
startup just to inspect a JSON document.

Drops the AddOpenApi/MapOpenApi wiring + StrongTypes.OpenApi.Microsoft
project reference from StrongTypes.Api, since OpenAPI is no longer being
exercised through it.

40 tests pass on both pipelines (1.1s — was 25s+ container boot).
