+++ src/StrongTypes.Api/Program.cs
@@ -1,16 +1,10 @@
 using Microsoft.EntityFrameworkCore;
 using StrongTypes.Api.Data;
 using StrongTypes.EfCore;
-using StrongTypes.OpenApi.Microsoft;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddControllers();
-builder.Services.AddOpenApi(options =>
-{
-    options.OpenApiVersion = Microsoft.OpenApi.OpenApiSpecVersion.OpenApi3_0;
-    options.AddStrongTypes();
-});
 builder.Services.AddDbContext<SqlServerDbContext>(options => options
     .UseSqlServer(builder.Configuration.GetConnectionString("SqlServer"))
     .UseStrongTypes());
@@ -21,7 +15,6 @@ builder.Services.AddDbContext<PostgreSqlDbContext>(options => options
 var app = builder.Build();
 
 app.MapControllers();
-app.MapOpenApi();
 
 app.Run();
