+++ src/StrongTypes.Api/StrongTypes.Api.csproj
@@ -6,7 +6,6 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <NoWarn>CS1591</NoWarn>
-    <InterceptorsNamespaces>$(InterceptorsNamespaces);Microsoft.AspNetCore.OpenApi.Generated</InterceptorsNamespaces>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="10.0.0" />
@@ -15,6 +14,5 @@
   <ItemGroup>
     <ProjectReference Include="..\StrongTypes\StrongTypes.csproj" />
     <ProjectReference Include="..\StrongTypes.EfCore\StrongTypes.EfCore.csproj" />
-    <ProjectReference Include="..\StrongTypes.OpenApi.Microsoft\StrongTypes.OpenApi.Microsoft.csproj" />
   </ItemGroup>
 </Project>
