+++ src/StrongTypes.OpenApi.IntegrationTests/StrongTypes.OpenApi.IntegrationTests.csproj
@@ -10,16 +10,15 @@
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
     <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
+    <RootNamespace>StrongTypes.OpenApi.IntegrationTests</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="10.0.0" />
     <PackageReference Include="Microsoft.Testing.Extensions.CodeCoverage" Version="17.14.2" />
-    <PackageReference Include="Testcontainers.MsSql" Version="4.0.0" />
-    <PackageReference Include="Testcontainers.PostgreSql" Version="4.0.0" />
     <PackageReference Include="xunit.v3" Version="3.2.2" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\StrongTypes.Api\StrongTypes.Api.csproj" />
-    <ProjectReference Include="..\StrongTypes.EfCore\StrongTypes.EfCore.csproj" />
+    <ProjectReference Include="..\StrongTypes.OpenApi.TestApi.Microsoft\StrongTypes.OpenApi.TestApi.Microsoft.csproj" />
+    <ProjectReference Include="..\StrongTypes.OpenApi.TestApi.Swashbuckle\StrongTypes.OpenApi.TestApi.Swashbuckle.csproj" />
   </ItemGroup>
 </Project>
