+++ src/StrongTypes.OpenApi.IntegrationTests/Tests/SwashbuckleOpenApiDocumentTests.cs
@@ -1,0 +1,15 @@
+using StrongTypes.OpenApi.IntegrationTests.Infrastructure;
+using Xunit;
+
+namespace StrongTypes.OpenApi.IntegrationTests.Tests;
+
+/// <summary>
+/// Runs the shared OpenAPI shape contract against the Swashbuckle test app,
+/// wired to <c>Swashbuckle.AspNetCore</c> (<c>AddSwaggerGen()</c>) +
+/// <c>Kalicz.StrongTypes.OpenApi.Swashbuckle</c>'s schema filters.
+/// </summary>
+public sealed class SwashbuckleOpenApiDocumentTests(SwashbuckleTestApiFactory factory)
+    : OpenApiDocumentTestsBase(factory.CreateClient()), IClassFixture<SwashbuckleTestApiFactory>
+{
+    protected override string DocumentUrl => "/swagger/v1/swagger.json";
+}
